回复功能验证

+++ src/main/java/life/majiang/comunity/comunity/controller/PublishController.java
@@ -28,7 +28,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String publish(@PathVariable(name = "id") Integer id,
+    public String publish(@PathVariable(name = "id") Long id,
                           Model model
     ) {
         Question question = questionMapper.selectByPrimaryKey(id);
@@ -40,7 +40,7 @@ public class PublishController {
 
     @PostMapping("/publish")
     public String doPublish(
-            @RequestParam(value = "id", required = false) Integer id,
+            @RequestParam(value = "id", required = false) Long id,
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag", required = false) String tag,
