+++ src/main/java/life/majiang/comunity/comunity/controller/QuestionController.java
@@ -19,7 +19,7 @@ public class QuestionController {
 
     @GetMapping("/question/{id}")
     public String profile(
-            @PathVariable(name = "id")Integer id,
+            @PathVariable(name = "id")Long id,
             Model model
     ){
         QuestionDto questionDto = questionService.getById(id);
