+++ src/main/resources/db/migration/V5__Create_comment_table.sql
@@ -1,9 +1,9 @@
 create table comment
 (
-	id BIGINT auto_increment primary key,
-	parent_id BIGINT not null,
+	id bigint auto_increment primary key,
+	parent_id bigint not null,
 	type int not null,
-	commentator int not null,
+	commentator bigint not null,
 	gmt_create bigint not null,
 	gmt_modified bigint not null,
 	like_count bigint default 0,
